requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aistudio-sdk==0.2.5
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
argcomplete==3.5.0
astor==0.8.1
astroid==2.13.5
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.15.0
backcall==0.2.0
bce-python-sdk==0.9.17
beautifulsoup4==4.12.3
bitsandbytes==0.43.2
black==24.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
codespell==2.3.0
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.21.0
decorator==5.1.1
decouple==0.0.7
Deprecated==1.2.14
dill==0.3.4
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
filelock==3.15.4
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
identify==2.6.0
idna==3.8
iniconfig==2.0.0
InstructorEmbedding==1.0.1
ipython==8.12.3
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
llama-cloud==0.0.15
-e git+https://github.com/run-llama/llama_index.git@cb6e9ddeb11132e9dcecb4823d70111328365f1a#egg=llama_index
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.3
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.3
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.3.0
llama-index-llms-openai==0.2.0
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.5.1
logger==1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
minijinja==2.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.12.2
mypy==0.991
mypy-extensions==1.0.0
narwhals==1.6.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
nodeenv==1.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
ollama==0.3.2
onnx==1.16.2
openai==1.43.0
opt-einsum==3.3.0
packaging==24.1
paddle2onnx==1.2.6
paddlefsl==1.1.0
-e git+https://github.com/PaddlePaddle/PaddleNLP.git@77480a071d9f0f45de1bbc7dbf3ec0327fa35838#egg=paddlenlp
paddlepaddle-gpu==2.6.1.post120
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipx==1.7.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.2.0
prettytable==3.10.2
prompt_toolkit==3.0.47
protobuf==5.28.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pylint==2.15.10
pyparsing==3.1.2
pypdf==4.3.1
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
rarfile==4.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rouge==1.0.1
rpds-py==0.20.0
ruff==0.0.292
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
striprtf==0.0.26
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenize-rt==6.0.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tool-helpers==0.1.1
torch==2.3.1
torchaudio==2.3.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
triton==2.3.1
typer==0.12.5
types-cffi==1.16.0.20240331
types-Deprecated==1.2.9.20240311
types-docutils==0.21.0.20240724
types-protobuf==4.25.0.20240417
types-pyOpenSSL==24.1.0.20240722
types-PyYAML==6.0.12.20240808
types-redis==4.5.5.0
types-requests==2.28.11.8
types-setuptools==67.1.0.0
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
userpath==1.9.2
uvicorn==0.30.3
virtualenv==20.26.3
visualdl==2.5.3
watchdog==4.0.2
watchfiles==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
